controllers/requirements.txt
sanic
sanic_cors
zlib
pillow
pyzbar
base45
cbor2